Create LongestArithmeticSubarray.cpp
Here the Arithmetic array means that consecutive elements should have an equal difference, for example, <9 7 5 3> or <3 3 3>.
So we need to find the longest subarray from an array having arithmetic array property.
Here pd is the variable used to keep track of the difference between the previous two elements and it is initialized with the difference between the first two elements of an array. Ans is the variable to hold maximum size and is initialized with 2, curr is var tracking the current difference between two elements and increases only when the same subarray.