Create indexOfRepeatedElement.cpp
Given an array of size N. The task is to find the first repeating element in the array of integers,i.e. an element that occurs more than once and whose index of the first occurrence is smallest.
Basic Idea:
Make an array idx that maintains the first occurence position and initalised with -1.
Take a var minidx to store the min no. of repeatition of an element.
minindx is initially -1 and updates with the position of a no. if it is repeating.